Implement advanced Find feature
Currently, the implementation of Find only supports searching for
keywords in the Internship's companyName.

This is quite limited as there are many other fields in Internship
that are of interest, and would be useful to filter by.

Let's
* Implement new predicates to test for keywords in other fields
* Implement 2 modes for find, withall or withany.
        * withall finds Internships that satisfy all listed predicates
        * withany finds Internships that satisfy at least 1 predicate

withall is useful when the search needs to be narrowed down further,
for example to find Internships whose locations are LOCAL and status
as ACCEPTED.

withany is useful when one may be looking for Internships
that are either REJECTED or have role including 'UI'.

Note: Within each prefix field specfied, the predicate still returns
true as long as any of the keywords are contained by the field. The
withall and withany mode only applies to predicates for separate fields.